Change GET decrypted secrets to unique URI
Currently retrieving the metadata and actual decrypted data for a
secret stored in Barbican uses the same URI, with only the Accept
header used to determine which response to return. This complicates
deployments with RBAC systems in front of Barbican (such as Repose -
http://openrepose.org/) or as middleware (such as EOM -
https://github.com/racker/eom). It also requires adding logic
like to the Barbican app when it is used to enforce RBAC.

This blueprint proposes using a unique URI to access decrypted secrets,
such as this: <host>/v1/secrets/<secret-UUID>/payload

The existing Accept-based mechanism for decrypting secrets will be
preserved to avoid breaking 'v1' API functionality, but will be marked
as 'deprecated' in API documentation.

Change-Id: Id654eb98e4fc2cfdf13d191258321964c9dfdc73
Implements: blueprint api-change-get-secrets-decrypted
APIImpact: Add alternate URI to decrypt secrets to ease RBAC policy
DocImpact: Existing means to decrypt secrets will be marked deprecated