Remove the tenant-secret association table
Currently all secrets stored in Barbican are indirectly associated with
projects via a tenant-secret association table. This association was
configured early in the project lifecycle, when it seemed that such an
association would aid in retrieving different components of grouped
secrets with different RBAC behaviors. However, recent cross-project
discussions have indicated that this feature would be better
implemented in a different way. Hence this association is adding an
additional join on selects with no benefit, and therefore should be
removed.

Change-Id: Ie53cf3c36afdd314e4a410acb27be1ba4f6ec71d