Content Types
The current implementation of Barbican is ambiguous to end users and
developers as to the encoding that should be used to input new secrets
and the encoding of secrets that are returned. This spec aims to
define the encoding specifications for each of the secret types and
define all of the secret types. The encoding specifications will allow
end users and developers to know how to handle the secrets.

Change-Id: I043f3487376c6b2183c4c381f9f823754917519f
Implements: blueprint content-types
DocImpact: The API will have new options for specifying the secret type
APIImpact: A new option will be added to specify secret type for storage