Add PUT Support For the Generic Containers Resource
The original intent of the 'generic' container type was to support
arbitrary secrets to be collected and referenced, similar to a file
system folder. Hence it makes sense for an already-created 'generic'
container to support changing this collection after the fact. This
blueprint details how this API would look.

Implements: api-containers-add-put
APIImpact: Add PUT action to the containers resource
DocImpact: Add docs for the new PUT action for containers resource
Change-Id: Ia2c9758436937096ff7d4c5f4ea89c302cc6983a