import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Remove stein from toc, there's no spec yet and zuul complains about an
empty toc.

Change-Id: Iede70fc477d0fe7b0b63336d1e120d5d5c23160d
Story: #2002586
Task: #24285